Button zones: Use xmin offset for button zones
When button zones are in use, the code attempts to calculate which
zone the finger is in.  The left and right edges of the zones are
calculated according the nths of pad_width, where n is the number of
zones.

These left and right sides are compared with `pos', and the zone is
the one where left<=pos<=right.

This would all be correct if pos were the touch's x coordinate
measured from the pad minimum.  However, inexplicably, pad_width/2 is
used instead of the pad minimum x coordinate.

The git history does not reveal an explanation for this.  I guess it
must have been empirical.

Using the xmin value makes the button zone code work on the two
laptops I have to test, here, one with xmin==0 and the other with
xmin=1200,xmax=4600.

Signed-off-by: Ian Jackson <ijackson@chiark.greenend.org.uk>
---
v2: This patch redone in the light of the fact that the origin
    is not always (0,0)